+++ e2e/react-start/basic/src/utils/cast.ts
@@ -1,0 +1,3 @@
+export function cast<T>(value: unknown): T {
+  return <T>value
+}