+++ packages/router-plugin/src/context.ts
@@ -1,0 +1,2 @@
+export { createRouterPluginContext } from './core/router-plugin-context'
+export type { RouterPluginContext } from './core/router-plugin-context'