+++ packages/router-cli/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @tanstack/router-cli
 
+## 1.166.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.41
+
+## 1.166.41
+
+### Patch Changes
+
+- Updated dependencies [[`056337e`](https://github.com/TanStack/router/commit/056337ef02ccbeaf45ec8533a227761be869e8f3)]:
+  - @tanstack/router-generator@1.166.40
+
+## 1.166.40
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.39
+
+## 1.166.39
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.38
+
+## 1.166.38
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.37
+
 ## 1.166.37
 
 ### Patch Changes