+++ packages/router-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tanstack/router-utils
 
+## 1.161.8
+
+### Patch Changes
+
+- Parse plain TypeScript files without JSX when a filename is available, preventing angle-bracket type assertions from being interpreted as JSX during route and Start import-protection transforms. ([#7342](https://github.com/TanStack/router/pull/7342))
+
 ## 1.161.7
 
 ### Patch Changes