+++ packages/solid-start/CHANGELOG.md
@@ -1,5 +1,128 @@
 # @tanstack/solid-start
 
+## 1.167.61
+
+### Patch Changes
+
+- Updated dependencies [[`35e88f0`](https://github.com/TanStack/router/commit/35e88f04996d71019a1868b7b06ecb4ddbc9df9e)]:
+  - @tanstack/start-client-core@1.168.2
+  - @tanstack/start-plugin-core@1.169.19
+  - @tanstack/start-server-core@1.167.30
+  - @tanstack/solid-router@1.169.2
+  - @tanstack/solid-start-client@1.166.47
+  - @tanstack/solid-start-server@1.166.51
+
+## 1.167.60
+
+### Patch Changes
+
+- Updated dependencies [[`056337e`](https://github.com/TanStack/router/commit/056337ef02ccbeaf45ec8533a227761be869e8f3)]:
+  - @tanstack/start-plugin-core@1.169.18
+
+## 1.167.59
+
+### Patch Changes
+
+- Updated dependencies [[`afa40ef`](https://github.com/TanStack/router/commit/afa40ef46c273c53646ba33f607fc618a07ffede)]:
+  - @tanstack/start-server-core@1.167.29
+  - @tanstack/solid-start-server@1.166.50
+  - @tanstack/start-plugin-core@1.169.17
+
+## 1.167.58
+
+### Patch Changes
+
+- Updated dependencies [[`709627f`](https://github.com/TanStack/router/commit/709627f3dbc6d97daa547a1401ef42a53bc4be32)]:
+  - @tanstack/start-server-core@1.167.28
+  - @tanstack/solid-start-server@1.166.49
+  - @tanstack/start-plugin-core@1.169.16
+
+## 1.167.57
+
+### Patch Changes
+
+- Add experimental HTTP 103 Early Hints support to Start server handlers. ([#7324](https://github.com/TanStack/router/pull/7324))
+
+- Updated dependencies [[`238ea4a`](https://github.com/TanStack/router/commit/238ea4a4998ab3a7fd528b317e1935766ac65df8)]:
+  - @tanstack/start-server-core@1.167.27
+  - @tanstack/solid-start-server@1.166.48
+  - @tanstack/start-plugin-core@1.169.15
+
+## 1.167.56
+
+### Patch Changes
+
+- Updated dependencies [[`96818b8`](https://github.com/TanStack/router/commit/96818b8ba5ead6f1f027094841330182aff415b2)]:
+  - @tanstack/start-plugin-core@1.169.14
+
+## 1.167.55
+
+### Patch Changes
+
+- Updated dependencies [[`ae453b7`](https://github.com/TanStack/router/commit/ae453b78624cac1b574f0d1efbfbf6ca03922c6c)]:
+  - @tanstack/start-plugin-core@1.169.13
+
+## 1.167.54
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/solid-router@1.169.1
+  - @tanstack/solid-start-client@1.166.46
+  - @tanstack/solid-start-server@1.166.47
+  - @tanstack/start-client-core@1.168.1
+  - @tanstack/start-plugin-core@1.169.12
+  - @tanstack/start-server-core@1.167.26
+
+## 1.167.53
+
+### Patch Changes
+
+- Updated dependencies [[`82b0613`](https://github.com/TanStack/router/commit/82b06132af776f74603ab27977cc277d6219a845)]:
+  - @tanstack/start-client-core@1.168.0
+  - @tanstack/solid-start-client@1.166.45
+  - @tanstack/solid-start-server@1.166.46
+  - @tanstack/start-plugin-core@1.169.11
+  - @tanstack/start-server-core@1.167.25
+
+## 1.167.52
+
+### Patch Changes
+
+- Updated dependencies [[`c4256c2`](https://github.com/TanStack/router/commit/c4256c2c857f392d2031cf87821e4c36a92d0382)]:
+  - @tanstack/start-plugin-core@1.169.10
+
+## 1.167.51
+
+### Patch Changes
+
+- Updated dependencies [[`761fcc0`](https://github.com/TanStack/router/commit/761fcc0c96dd96721b533a1fd9e2c972f222ef94)]:
+  - @tanstack/start-plugin-core@1.169.9
+
+## 1.167.50
+
+### Patch Changes
+
+- Updated dependencies [[`c992495`](https://github.com/TanStack/router/commit/c992495bf4010ff4c3597bb1f3b1ba02594e857e)]:
+  - @tanstack/solid-router@1.169.0
+  - @tanstack/solid-start-client@1.166.44
+  - @tanstack/solid-start-server@1.166.45
+  - @tanstack/start-client-core@1.167.22
+  - @tanstack/start-plugin-core@1.169.8
+  - @tanstack/start-server-core@1.167.24
+
+## 1.167.49
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/solid-router@1.168.26
+  - @tanstack/solid-start-client@1.166.43
+  - @tanstack/solid-start-server@1.166.44
+  - @tanstack/start-client-core@1.167.21
+  - @tanstack/start-plugin-core@1.169.7
+  - @tanstack/start-server-core@1.167.23
+
 ## 1.167.48
 
 ### Patch Changes